requirements.txt
python-dateutil==2.9.0.post0
python-decouple==3.8
six==1.17.0